Hacking a workaround for #94
This is DEFINITELY not a solution, but the last time I coded obj-c was in school, 11 years ago. I'm definitely not willing to understand in which scenarios the completionHandler is called more than once. I hacked this solution with a rough idea and ChatGPT's guidance on how to do almost anything in this weird language.

WORKAROUND DISCLAIMER (if it isn't clear enough from the code comment): if you repeat the same notification body, it will probably have the same notId (it had in my test cases), and thus it will fail to call the completion handler, causing a warning. However, I guess that's better than a full application crash lol Not sure if that could cause a memory leak or stuff like that... also, not sure if this arbitrarily long NSMutableDictionary could grow terribly large and cause issues; I hope your app doesn't get too many notifications in its lifetime.